security: clear CodeQL alerts (CI token least-privilege + test overflow cast)
- Add a least-privilege `permissions: contents: read` to the CI workflow. The
  jobs only check out and build/test, so the default broad GITHUB_TOKEN scope
  isn't needed. Clears the three actions/missing-workflow-permissions alerts.
- Cast to size_t before the multiply in test_deswizzle.c so the tiled-size
  computation runs in 64-bit instead of 32-bit then widened. Clears the
  cpp/integer-multiplication-cast-to-long alert. Test only and the values are
  tiny, but the pattern is worth fixing.

No change to the shipped library or the privileged helper, both already had
zero CodeQL alerts.

+++ tests/test_deswizzle.c
@@ -145,7 +145,7 @@ static void test_deswizzle_nvidia_x_tiled_roundtrip(void) {
 
     uint8_t *linear = calloc(1, stride * h);
     /* Tiled size: tiles_x tiles × (tile_w_bytes × tile_h) bytes */
-    size_t tiled_size = tiles_x * tile_w_bytes * tile_h;
+    size_t tiled_size = (size_t)tiles_x * tile_w_bytes * tile_h;
     uint8_t *tiled = calloc(1, tiled_size);
     uint8_t *result = calloc(1, stride * h);
     TEST_ASSERT(linear && tiled && result);
